Building: remove the 4m and 8m split images from the docs
Flashing the generated 8M image is dangerous. It's not usable on real
systems.

In order to have a system that can be easily flashed externally, there
is the x230-flash make target. The 8M chip though should only be touched
externally for unlocking the IFD (and optionally cleaning the ME).