index.md: show both fbwhiptail(high res)/whiptail(low res) main menu
This shows the power of using fbwhiptail/fbwhiptail to drive menus under Heads.

Whiptail is command-line utility used in shell scripts to create text-based user interfaces (TUIs) in a terminal (console).
FBwhiptail is a whiptail wrapper, aimed at driving a reduced feature set of whiptail but for high resolution Frame Buffer (FB), intruduced by Purism.

By using Whiptail in the shell scripts, we are able to drive both high resolution graphics for laptops and workstations into coreboot initialized simple framebuffer, picked up by efifb linux kernel driver, which permits us to kexec into a full featured linux kernel, at the condition that the initramfs of that next kernel also provides efifb, which is currently being deprecated.
As of now, neither simplefb, nor simpledrm permits the first kernel context (Heads) to be passed to the next kernel to provide early console access for the user to type its LUKS decryption passphrase at systemd prompt if efifb is not provided in the next kernel's initramfs. This is the case with Ubuntu 20.04+, but not Debian nor Fedora as far as I know. Ubuntu is not going back into their decision of readding efifb into their distributed initramfs as can be seen in issue https://github.com/linuxboot/heads/issues/1641
Mitigation to this for Heads users wanting to boot into OSes that deprecates LinuxBoot based implementations is to enable TPM Disk Unlock Key, so that the decryption key passphrase is passed into a cpio, not asked by systemd which doesn't have a fully functional console yet; only after rootfs is fully accessible and full GPU drivers replaces simpledrm/simplefb.

Let's remember that neither simpledrm nor simplefb can expose their framebuffer addresses to be passed to the next kernel, same applying to fully fledged linux kernel drivers for i915drm/i915 which needs insecure parameters to be passed.
So for the future of linuxboot implementations desiring to provide a great UX, the issue needs to be tackled from coreboot to provide simpledrm compatible framebuffer, and for kexec to be able to pass that framebuffer to next kernel.

Signed-off-by: Thierry Laurion <insurgo@riseup.net>